backend/requirements.txt
Django==5.2
djangorestframework==3.14.0
numpy
pandas
scikit-learn
django-cors-headers
